requirements.txt
# Core LLM clients
openai>=1.50.0,<2
anthropic>=0.39.0,<1

# Orchestration
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Search APIs 
exa-py>=1.0.0

# HTTP / async
httpx>=0.27.2
requests>=2.32.3
aiohttp>=3.9.0
tenacity>=8.5.0

# HTML parsing (for fallback)
lxml>=5.3.0

# Validation & config
pydantic>=2.10.0
pydantic-settings>=2.6.0 # Better for .env handling
python-dotenv>=1.0.1

tiktoken>=0.7.0

# Date parsing (for publish date metadata)
python-dateutil>=2.9.0

# UI
streamlit>=1.38.0
